I am checking in Gregory Malecha's Coq plugin for reifying internal Coq representation into a Coq datatype.  (See template-coq/theories/Ast.v for the Ast of terms and programs. "Programs" are hereditarily extracted, containing everything needed to evaluate the last term in the program.)
Gregory's reification does no "extraction" or optimization, it is
intended to be a (correct) reification of coq internal syntax.

I've checked in only sources (ocaml sources, coq sources and examples,
which are also coq sources).  You will need to "make" in the
template-coq directory.

Later today I will check in (work in progress) evaluator of programs.