I've reorganized the non-deterministic step relation and added a functional cbv evaluator.
File evalShared.v is common to both evaluators.  It needs to be
compiled to run either of the evaluators.  (coqc evalShared.v)

File eval.v is the non-deterministic step relation from before.

File functnlEval.v is the new functional evaluator.  It has a clock to
make it total.  If the returned answer is (0, someTerm) then it ran
out of time, and "someTerm" is probably not in whnf.

Both evaluator files have some test examples.  The functional
evaluator can easily compute fib 20 (3 secs).  (This is not a good
test example, because it is hard to tell if the answer, in Peano
numbers, is correct.)

Still, nothing is proved, and the evaluators are only correct under the
assumption that the program being evaluated is accepted by Coq's
typechecker.

Clearly we should prove something like "the functional evaluator
results are in the transitive closure of the non-deterministic step
relation".  There are some subtleties about termination with an
incomplete evaluation vs getting stuck.  Also behaviour on
non-type-correct terms may differ.

Randy