Adding this file to make sure it is in the svn log.
The useful files in this commit are:

RandyPrelude.v, evalShared.v, wndEval.v, wNorm.v

which must be compiled in that order.

test_wndEval.v and wcbvEval.v are in progress at the moment.

The main definitions (that a reader must agree with) are Term and Crct
(in evalShared.v), wndEval (in wndEval.v) and WNorm (in wNorm.v).  The
main theorem is no_wndStep_wNorm (in wNorm.v) saying that if a correct
program cannot step using wndEval, then the body term is in weak
normal form.

I will send a more detailed email later.  These current definitions
still need work.