In a previous commit I added a _relation_ of weak-call-by-value evaluation (in file wcbvEval.v) and a proof that WcbvEval is contained in the reflexive-transitive closure of small-step non-deterministic reduction (wndEval).
I did that because if we want to prove correctness of cbv CPS
translation, we will use this relation as a correctness specification.

This commit includes a weak-call-by-value evaluation _function_ (in
file wcbvEval.v) which can be executed on quoted programs, and a proof
that the evaluation relation and function have the same extension.

I'm working on some other things in the files too, but they are
commented out, and "make" checks the whole development so far.

Randy
--