In our last Skype meeting we said that (for now) we focus on a type-stripped language.  This commit provides every result for the stripped language that I had proved for the typed language.  See files stripTerm.v stripProgram.v stripWndEval.v stripWcbvEval.v stripEvalCommute.v.
Until now the AST for Malecha's quoted term representation looked
like:

Inductive term : Type :=
| tRel       : nat -> term
| tSort      : sort -> term
   ...
| tUnknown : string -> term.

If you tried to Quote a CoFixpoint definition (for example), you got
tUnknown (since we are not currently handling CoFicpoint).

I don't see the point of our correctness reasoning carrying around
tUnknown, so I changed the Quoting plugin: it only returns quoted
forms of programs we intend to handle.  If you try to quote something
we don't (yet) handle, the plugin fails by raising an exception
(caught by Coq and displayed as as "anomaly" error) rather than
returning tUnknown.

In fact I was using tUnknown as a "bottom" element of terms.  I knew
this was a bad idea, and I finally got around to cleaning this up.

Now I use an option type for a bottom element of term, rather than
tUnknown.  The small step weak reduction relation still has the same
type as before (since it is a relation and doesn't have to be total):

     wndEval: environ -> Term -> Term -> Prop

But the big step evaluation _function_ now has type:

    wcbvEval: (fuel:nat) -> environ -> Term -> option Term.

You might wonder how wcbvEval can fail.  For example, the
[instantiation] function cannot fail, and is correct whenever the term
being implanted is locally closed (has no unbound deBruijn indices).
Thus locally closed pure lambda terms are correctly beta-evaluated
until the fuel runs out, without any way to fail.

But Case and Fixpoint introduce possibilities for failure. For example
a Case may not have enough branches for the constructor in the match,
and a mutual Fixpoint may not bind the right number of variables (the
number of functions in the mutual package).

Of course well-typed terms should not fail in either of these ways,
but I know of no way to check for these failure modes statically
(other than typechecking, which is not our plan).


In an 18 Jan commit I claimed:

   (1) Proved: when weak small-step reduction (wndEval in file
   wndEval.v) cannot step a program, then that program is in weak
   normal form (WNorm in file wNorm.v).

This is no longer true with the strict notion of terms without
tUnknown, as Case and Fixpoint syntactic redexes may not actually step
for the reasons mentioned above.  The converse of (1) is still true.

Randy
--