Comparing non-deterministic small-step weak reduction (wndEval in stripWndEval.v) with the big-step weak evaluation _relation_ (WcbvEval in stripWcbvEval.v) I previously proved the theorem:
Lemma WcbvEval_wndEvalRTC:
  forall (p:environ),
    (forall t s, WcbvEval p t s -> wndEvalRTC p t s) /\
    (forall ts ss, WcbvEvals p ts ss -> wndEvalsRTC p ts ss).

Big-step evaluation is in the REFLEXIVE-transitive closure of
small-step reduction.  We need the reflexive closure here because,
e.g. \x.M weakly-evaluates to itself, but has no weak reduction step.

Still, this theorem is unsatisfactory because we don't mean to allow
unrestricted use of reflexivity; only where WcbvEval is reflexive.  I
am now committing the following stronger theorem:

Lemma WcbvEval_wndEvalTC:
  forall (p:environ),
    (forall t s, WcbvEval p t s -> t <> s -> wndEvalTC p t s) /\
    (forall ts ss, WcbvEvals p ts ss -> ts <> ss ->
                    wndEvalsTC p ts ss).

Where [wndEvalTC] is the transitive closure of wndEval.  The proof of
this theorem is more picky than the proof of the weaker theorem, and
uses the decidability result on Term equality that I checked in
yesterday.


As an aside, notice that wndEval is not anti-reflexive, since
\Omega=(\x.xx)(\x.xx) reduces to itself in one step.  \Omega is not
typable in CiC. but our quoted Coq theory with no typechecking can't
prove that.

As another aside, it occurred to me that we might improve the
confidence of our use of Malecha's quoting plugin by having it
cryptographically sign the quoted terms.  This is an idea for the
distant future.