All of the "common" and "L1" directories typechect with Malecha's main branch.
The translation between Malecha's Ast and my L1 version of the Ast
(look for the inductive definitions of Term and Program) use an
exception monad (e.g. to remove Malecha's term constructor
[tUnknown]), so actual test evaluations (e.g. in test_wcbvEval.v) must
be aware of the monad.  However most of the reasoning about L1 just
uses my Term and Program Asts, so is innocent of the monad.