Added support for let and match to the simple CPS translation. For the most part, this wasn't much more difficult than what I had before.
I also changed the representation of both the source and the
target language to use lists where appropriate instead of 
defining, for instance, my own inductive definition for 
lists of expressions.  This required that I write down my
own induction principles explicitly (instead of generating
them with a Scheme) but meant that I didn't have to re-prove
things about associativity of lists, lengths of lists, etc.