There is a proof of
Theorem wcbvEval_types_irrelevant:
  forall p, L1.program.WFaEnv p ->
    (forall t s, L1.wcbvEval.WcbvEval p t s -> L1.term.WFapp t ->
          WcbvEval (stripEnv p) (strip t) (strip s)) /\
    (forall ts ss, L1.wcbvEval.WcbvEvals p ts ss -> L1.term.WFapps ts ->
          WcbvEvals (stripEnv p) (strips ts) (strips ss)).

in certicoq/examples/RandyWorking/L2_typeStrippedL1/stripEvalCommute.v.

This states that stripping type fields from L1 commutes with cbv
evaluation; i.e. that L1 and L2 are equivalent for cbv evaluation on
well formed programs.

BTW, this was not as trivial as it sounds.  You will notice that L2
still has a type constructor [Cast], although this carries no type
information.  Some time ago I had removed [Cast] from L1 (since its
only reduction rule just forgets the cast), but Andrew wasn't happy
with that.  So I put [Cast] back in L1, and found I couldn't remove it
from L2 because the interaction between casts and application to lists
(where the function position cannot be an application) was too
difficult to reason about, since stripping a cast can make a
well-formed application into an ill-formed one, and fixing it up again
completely changes the term structure.  The spurious [Cast] of L2 can
be removed in the translation to L3 (which has unary applications).  I
will have to revisit the translation to L3.

There is a theory of "spine lambda terms" (Cervesato and Pfenning)
that handles application to lists, but it syntactically distinguishes
between terms and spines.  Maybe the Coq kernel should use this spine
representation, but it doesn't, so I didn't use it either.