There is a proof of
Lemma WcbvEval_wndEvalRTC:
  forall (p:environ), WFaEnv p ->
    (forall t s, WcbvEval p t s -> WFapp t -> wndEvalRTC p t s) /\
    (forall ts ss, WcbvEvals p ts ss -> WFapps ts -> wndEvalsRTC p ts ss).

in certicoq/examples/RandyWorking/L2_typeStrippedL1/wcbvEval.v.

This states that (for terms that have only well-formed applications),
big-step cbv evaluation is contained in (sound for) the
reflexive-transitive closure of small-step evaluation.

Although for the moment we are not using small-step evaluation as our
official specification, I was interested in this lemma because I find
small-step evaluation more clearly correct than big-step cbv.

There are also some new sanity lemmas in wNorm.v.

- If a program is well-formed and cbv evaluates to term t, then t is
  in weak normal form.  This is analogous to Greg's lemma
  eval_yields_value.

- Values cbv-eval to themselves: WcbvEval p t s -> WcbvEval p s s

- values can take no small-steps: WNorm t -> no_wnd_step p t.


One congruence rule of wndEval (small-step) is unusual:

| sAppFn: forall (t r arg:Term) (args:Terms),
              wndEval p t r ->
              wndEval p (mkApp t (tcons arg args))
                        (mkApp r (tcons arg args))

mkApp is the pseudo constructor that preserves the invariant that
there are no nested applications. It is clearly needed on the RHS of
this rule.  One expects the rule to be

| sAppFn: forall (t r arg:Term) (args:Terms),
              wndEval p t r ->
              wndEval p (TApp t (tcons arg args))
                        (mkApp r (tcons arg args))

but I need the previous rule for purposes of induction.  There is a
new file awndEval.v (for "alternative wndEval") that proves the two
inductive definitions are the same for well-formed terms.