Committing the version of cps I've been working on.  This is not yet complete, and not yet integrated into the main development (L1, L2, L3, ...).
This new cps directory has files

- expression.v 
  Greg's direct expression language, substitution and evaluation for
  that language.  Some sanity checks (evaluation is idempotent and
  single-valued), and some test examples.

- cpstrans.v
  Greg's cps language and cps translation.  This needs no substitution
  on the cps language, so can occur early in the development.  Some
  sanity checks (e.g. the range of cps translation is
  continuation-closed).

- k_deBruijn.v
  Shifting and substitution for continuation variables, and lemmas
  about these operations.

- u_deBruijn.v
  Shifting and substitution for user variables, and lemmas about
  these operations.

- cpseval.v
  Evaluation and reduction of the cps language.  Test examples.

I became somewhat obsessed with this part of the project.  Some months
ago Greg asked me to look into a problem concerning the relationship
between evaluation and reduction of the cps language.  This problem
hung on technical lemmas about shifting and substitution.

The first thing I saw was that Greg (possibly following Xavier) had
used an "optimised" definition of shifting: instead of shifting by one
when going under each binder, save up the shifts (in an auxiliary
argument to the shift function) and shift all at once when the target
variable is reached.  All the literature says optimised shifting and
reduction is much harder to reason about, so I defined the more
standard unoptimised shifting and reduction for both k-variables and
u-variables of the cps language.  (That's four shifts and four
substitutions: optimised and unoptimised for k-variables and
u-variables.)

Since both evaluation and reduction only mention substitution for
index 0, the job is "just" to show that optimised and unoptimised
k-substitution are equivalent, and similarly for u-substitution.
Although I easily proved this for a cut-down language (following a
paper by Nipkow), I consistently failed to prove it for the full cps
language.  

Indeed, I did not understand the treatment of indices in what Greg had
written, so I tried to clarify it for myself by proving sanity lemmas,
and by computing examples. 

I found Greg's presentation of expressions and cps terms using lists
(and the associated boilerplate for unwinding mutual dependencies) to
be hard to reason about, so I rewrote all the mutual inductive
definitions to use straightforward mutual induction.  This saved many
lines of code and made the development easier for me to understand.
The more straightforward Fixpoints defined over these mutual
inductions also compute better in Coq.  I was finally able to prove
that cps translations are k-closed, which had eluded me for several
days.  (This was the hardest thing to prove so far, since some of the
substructures (e.g. branches) are not themselves k-closed, but are
only created under some binders.)

In order to validate the definitions of cps translation and
evaluation, I looked for examples to compute. Greg's original file had
a few examples to test cps translation. but no end-to-end examples of
translation of evaluation.  I've been working on examples for the last
few days. In relation to Matthieu and Greg's recent email, there were
no tests of Fixpoints at all, and I'm working on that now.

There are a few other things to note.  Evaluation of Greg's direct
language only works for closed expressions, and similarly for
evaluation of the cps language.  For evaluation (i.e. weak-head
reduction) of closed terms, no shifting at all is needed in the
definition of substitution.  So I've been experimenting with no-shift
substitutions.  The only reason to use substitutions that are correct
for open terms is if the compiler itself will do reduction under
binders as part of its optimization. (Greg? Andrew?)

Finally a minor point, I (temporarily) removed an optimization of
Greg's in the cps translation where data constructors that are fully
applied to values were treated specially.

Best,
Randy