L3.term.Term has a new constructor, TProof.
TProof is a proof that wasn't removed by the translation from L0 to
L1.

TAx (which has been a L3.term.Term constructor for some time) is an
axiom that is not a proof.

The whole project compiles, but I had to make very minor changes in
L3_to_L4.v: new dummy case in function [trans].

Aside from compiling, this version is not well tested, so let me know
of any problems you find.

--Randy