Added a style document to guide refractoring and future decisions.
Adapted from https://github.com/HoTT/HoTT/blob/master/STYLE.md
These guidelines are just a starting point, and may not be optimal.
Please feel free to edit and discuss.

Once/if we have a consensus, we can start the process of refractoring, and stick to the agreed guidelines for future developments.