new correctness property of phases, proof of its composition.
The CerticoqLanguage2 class requires the following : a type of values, which may be the same as the type of terms, well formedness predicate on terms, a big step eval relation from terms to values, a function to observe the head constructor of a value (if any), a function to observe constructor arguments of a value (if any).

The class CerticoqTranslationCorrect2 requires a proof of the following
1) the translation preserves well-formedness
2)  ∀ (s:Src) (sv: SrcValue),
    wf s
    -> (s ⇓ sv)
    -> ∃ (dv: DstValue), (translate Src Dst s) ⇓ (Ret dv) ∧  sv ~ dv.

~ is coinductively defined using the observation functions mentioned above.

The instances of this class compose -- a proof of correctness of Li to Lk via Lj can be automatically obtained by proofs fof Li to Lj and Lj to Lk.
See common.certiClasses.composeCerticoqTranslationCorrect2