made L4 to L4a faster : 40s to 0.002s (Eval compute)
Eval vm_compute was fast even before the change (0.003s), and is still so.
Used another implementation debruijn to named conversion which has N log N complexity (linear if +,- on numbers is constant time).
However, the main bottleneck was somewhere else -- in SquiggleEq term representation!

It was using a "bundled"[1, Sec. 3] approach where each (sub)term has the whole description of the language (including recursive functions denoting the signature of operators, decidable equality of operators) as an implicit argument.
Now, it is almost totally unbundled. The type of operators is the only implicit argument in (sub)terms.

The SquiggleEq LIBRARY NEEDS TO BE REBUILT. Execute make -j 4 at the root of the repository (parent of theories/).

[1]: Spitters, Bas, and Eelis Van Der Weegen. “Type Classes for Mathematics in Type Theory.” MSCS 21, no. Special Issue 04 (2011): 795–825. doi:10.1017/S0960129511000119.
http://arxiv.org/pdf/1102.1323v1.pdf