changed the representation of Fix in L4a,L5, L5a
Not using the projection based representation of mutual fixpoints.
Now, there are separate bound variables for each member in a mutual fixpoint block.
Had to update the CPS translation and its tricky correctness proof.
The bodies of Fix, which must be lambdas, are now translated using the value translation (cps_cvt_val).
The replacement of the projection rule in eval_c is:

| eval_FixApp_c : forall lbt i k arg v bt,
    let len := Datatypes.length lbt in
    let pinds := seq 0 len in
    let sub := map (Fix_c' lbt) pinds in
    select i lbt = Some bt ->
    eval_c (Call_c (apply_bterm bt sub) k arg) v ->
    eval_c (Call_c (Fix_c' lbt i) k arg) v.

Now, these languages (L4a-L5a) are similar to the rest in terms of how fix is represented.
The proof of L4 to L4a correctness and L5a to L6 should be easier now.

Fixed the (L4) files broken due to broken imports due to the namespaces changes in Zoe's r7636, and added them back in _Coqproject
Everything (in the _CoqProject(s)) compiles on my machine when I do make clean; make -j8 -k at the root of the repo.