fixed the bug pointed out by Olivier.
In L5 to L5a, the order of branches in matches were reversed (fold left vs fold right).
This may not be a bug (each branch carries constructor information), but the same problem also caused
mutual fixpoints to be reversed, which was most likely a bug.

I am hoping that we can eventually get rid of L5a -- thus I am not formalizing the correctness of L5 to L5a for now.