added the Gcd Function test case.
The last line prints the result of compiling Gcd_terminate to L3.

2 Observations:

1) The environment is too big, perhaps because it contains even the
definitions that were used in the erased proof. The whole result of compilation runs into 50K lines (using Coq's Print).
This makes it very important to prune the environment in L3 to L4.
Perhaps Certicoq should take as input the functions that the user is interested in compiling, and prune everything
in the environment that is not needed to get those functions to be well-defined. This is what the current
extractor does.

2) After projecting out just the definition of Gcd_terminate, we get 900 lines in L3.
When pretty printed, this may result in a much shorter term, but the size is still concerning.