defined L4.eval polymorphically over a weak interface.
This interface does not make any assumption about the style of variable
bindings : generic (Squiggle style) or concrete, de-bruijn or named, or...
Thus, it should be possible to instantiate it with both L4 and L4a.
Furthermore, the parametricity plugin should be able to prove that the
eval function behaves the same way in L4 and L4a.
The plugin would, however, need a proof that the translation from L4 to
L4a is a homomorphism w.r.t this interface. This boils down to proving that substitution
commutes with L4 to L4a translation.