added the declarations generated by the parametricity plugin for 8.6
In Coq 8.5, the plugin fails to create these declarations.
This failure is caused by a "bug" in Coq 8.5:
https://github.com/mlasson/paramcoq/issues/5

Marc Lasson had submitted a bugfix, but only the trunk and the v8.6
branches of Coq contain the fix.
If/when we move to Coq 8.6, instead of manually adding these
declarations, we can just invoke the plugin.
As of today, all the dependencies of Certicoq (Template-coq, Extlib,
SquiggeEq) fail to compile in Coq 8.6 (tip of the 8.6 branch at
github:coq/coq)