SquiggleEq now has a theory of generic de-bruijn (DB) terms.
This theory includes a simple de-bruijn substitution function (no
reduction of variable-indices), the complex substitution used in L4, and
a proof relating the two. It also has a function to convert from generic
DB to generic named terms. There is a proof that substitution commutes
with this translation.

Although these  proofs  are complex and go thru several intermediate
lemmas, it was much easier to do these proofs in the generic setting
because are no unnecessary details (e.g. proof by induction on generic
terms results in 3 cases, instead of the 14 in L4.expression.exp).

The de-bruijn substitition function doesn't yet perform lifting, which
is not needed for defining eval because eval only substitutes with
closed terms.

The translation from L4 to L4a will pass through an instantiation of
these generic DB terms and use the above-mentioned generic function for
DB to named conversion.