renamed the Class WellFormed to GoodTerm, as discussed last Wed.
WellFormed has connotations of decidability. The GoodTerm predicate on
L1g will not be decidable.

Class GoodTerm (Term: Type) := goodTerm : Term  -> Prop.

This class says that the type Term comes with a predicate `goodTerm`,
which characterizes the set of terms of that language that are
guaranteed to be correctly compiled to the next stage.

This is used in the following class which declares an end to end
correctness property that we want to prove:

Class CerticoqTranslationCorrect2
  `{CerticoqLanguage2 Src SrcValue}
  `{CerticoqLanguage2 Dst DstValue}
  `{CerticoqTranslation Src Dst} :=
{
  certiGoodPres2 : goodPreserving Src Dst;
  obsePres : obsPreserving Src Dst;
}.

where  obsPreserving  is defined as

Definition obsPreserving
  `{CerticoqTranslation Src Dst}
   `{BigStepHetero Src SrcValue} `{BigStepHetero Dst DstValue}
`{GoodTerm Src}
   `{ObserveHead SrcValue} `{ObserveSubterms SrcValue}
   `{ObserveHead DstValue} `{ObserveSubterms DstValue}
  :=
   ∀ (s:Src) (sv: SrcValue),
    goodTerm s (* Randy, please note this hypothesis *)
    -> (s ⇓ sv)
    -> ∃ (dv: DstValue), (translate Src Dst s) ⇓ (Ret dv) ∧  sv ⊑ dv.

For more details on these classes, please look at
theories/common/certiClasses.v

I suggest that every language should soon declare their instance of
GoodTerm. We should start from the later phases, because the goodness of
L_i should be strong enough to imply that the translation from L_i to
L_{i+1} preserves goodness.
The instance of L1g should be strong enough to
imply goodness preservation throughout the compilation chain.