formalized the end-to-end correctness property suggested by Greg.
The only change is in the definition of obsLe, the relation between (observations on)
values of Src and Dest. The new definition is based on
questions of the form "Are you the nth constructor of the inductive i".
A value can respond "yes" to many such questions.
Also, now a value cannot be asked how many subterms it has.

Proved that the new correctness property composes.
These new definitions and proofs are in theories/common/certiClasses2.v