A new definition of the parametric correctness property
This experimental definition tries harder to allow the possibility of
adding support for coinductive types (+ cofix) in the future.
The main idea is to more uniformly allow computation before observations.
This is especially important for subterms of coinductive constructors.

The new definition is in certiClasses3.v.
This file is not in use, and may never be in use.