Handle TProof correcly in L3/L4.
Add an App_e Prf_e a -> Prf_e reduction rule accordingly.
Changed L2k to translate TInd to a TProof, as these types
can be applied and we don't want to get reduction stuck because
of that. For TSort and TProd both TAx and TProof could be fine
but that would mean leaving axioms in later phases, for inert type
arguments, which does not seem necessary.