I have added a requested "well-formedness" check to L1g, L2, L2k, L2_5 and L3. Namely crctDs (a mutual relation with crctTerm, crctEnv, crctTerms, and crctBs) checks that the definitions of (mutually defined) fixpoints are actually lambda abstractions.
I order to do this I have cleaned up what was a minor hack.  I used to
have [Defs] (the "list" of mutually defined fixpoint definition
bodies) do double duty as the "list" of Case branches (just a hack to
package a term and a nat together).  I have now introduced a new
"list" package [Brs] for case branches, and [Defs] is only used for
fixpoint definitions.

This changes the L3 AST, so L3-to-L3eta-to-L4 translation no longer
checks (sorry Mathieu).

BTW, why can't the L3-to-L4 translation check this well-formedness
condition locally, and simply refuse to translate if it is not
satisfied, as indeed L1g, L2, L2k, L2_5 and L3 have their own reasons
to fail to translate.