simplified the linking illustration.
Now, instead of having another compiler comp2, we consider
the case where the argument (say ta) of application in the target language
is produced some by some arbitrary way (e.g. manually).

For the correctness of the target application, we still need that
ta observationally greater than what the original compiler produced.
Ideally, we would want it to be greater than the corresponding Gallina term.