Optimized by removing [inductive] argument from TCase constructor at language L2.
This argument is not used in comptation, and the proof that the
translation from L1g to L2 is correct still works.

L3_to_L3_eta now fails to check because it tries to check that the
inductive type of a case and its discriminee are the same, but this
has nothing to do with computation.  Also, some of this check is
Admitted, so I didn't try to fix it.