proved: eval_n commutes with L4 (concrete DB) to L4_1 (abstract DB)
For L4, the statement uses L4.expression.eval_ns instead of
L4.expression.eval_n.

The changes correspond mainly to the changes in L4.expression.eval
that I mentioned in my last email to the certicoq list, and recap below.

It remains to show that L4.expression.eval -> L4.expression.eval_ns,
which should be if the following changes are made to
L4.expression.eval:

1) remove:
| eval_Ax_e s : eval (Ax_e s) (Ax_e s)

2) change eval_ProofApp_e
from:
| eval_ProofApp_e : forall f a,
    eval f Prf_e ->
		eval (App_e f a) Prf_e
to:
| eval_ProofApp_e : forall f a,
    eval f Prf_e ->
    eval a av ->
		eval (App_e f a) Prf_e

3) change eval_Match_e
from:
| eval_Match_e: forall e p bs d vs e' v,
    eval e (Con_e d vs) ->
		let vs' := exps_skipn p vs in (* remove this. Randy is already deleting parameters *)
		find_branch d (exps_length vs') bs = Some e' ->
		eval (sbst_list e' vs') v ->
		eval (Match_e e p bs) v
to:
| eval_Match_e: forall e p bs d vs e' v,
    eval e (Con_e d vs) ->
		find_branch d (exps_length vs) bs = Some e' ->
		eval (sbst_list e' vs) v ->
		eval (Match_e e p bs) v