Update certicoq after move of names/inductives to the BasicAst library
This avoids importing the original Ast of Template-Coq terms when
working with languages defined down the line of the compiler which just
rely on the definitions of names/inductives/sorts etc and not the
original term structure. Required quite a few renamings, from Ast to
BasicAst: I may have missed some.