Update INSTALL.md
Update installation instructions to include the hidden dependencies on both Inria's repositories and in clang and libgmp-dev (also removed the pinning of submodules as it breaks the process). Also included the current and arguably easiest of installing coq-certicoq by just using Inria's repositories without even having to clone the repo.